// src/enums/plateVals.ts
export enum PlateVals {
  RedKG = 25,
  BlueKG = 20,
  YellowKG = 15,
  GreenKG = 10,
  RedLB = 55,
  BlueLB = 45,
  YellowLB = 35,
  GreenLB = 25,
  //all ratios are x to barlength
  //IWF
  IWFBumperHeight = 0.204545454545455,

  IWFRedKGWidth = 0.026363636363636,
  IWFBlueKGWidth = 0.023181818181818,
  IWFYellowKGWidth = 0.017727272727273,
  IWFGreenKGWidth = 0.015909090909091,

  IWFRedLBWidth = 0.028181818181818,
  IWFBlueLBWidth = 0.022272727272727,
  IWFYellowLBWidth = 0.018181818181818,
  IWFGreenLBWidth = 0.015454545454545,

  //IPF
  IPFRedKGWidth = 0.01227272727,
  IPFBlueKGWidth = 0.009090909091,
  IPFYellowKGWidth = 0.008636363636,
  IPFGreenKGWidth = 0.008636363636,

  IPFRedHeight = 0.2045454545,
  IPFBlueHeight = 0.2045454545,
  IPFYellowHeight = 0.1818181818,
  IPFGreenHeight = 0.1477272727,

  IPFRedLBWidth = 0.01227272727,
  IPFBlueLBWidth = 0.01,
  IPFYellowLBWidth = 0.009545454545,
  IPFGreenLBWidth = 0.01045454545,
}